setRootTemplate when CarPlay starts
This commit checks if the SwiftFlutterCarplayPlugin.rootTemplate is nil
If it is nil, that means the Flutter app hasn't start up. If we end up
in this situation, we want to manually call setRootTemplate to boot up
the Flutter app.

We need to do this since this package lives in the Flutter world. If
Flutter doesn't boot up then the CarPlay app will not work by showing
a blank screen on the device.