Check size of template history before removing last element
For some reason (probably a race condition when pushing a new screen and then popping it right away) the template history sometimes only contains one element (the root template). Removing the last element in this case would lead to a broken UI, so we must check if there are at least two elements.